239 - When debugging the REPL window, breakpoints on the first code line of a file are not hit 	Tweak workaround for IronPython bug – we need to always deliver the handle_line event for line 1 of modules, not just when we’re in breaking for attach mode.
233 - .NET documentation not showing up
	Add support for introspecting .NET objects in the REPL when running on IronPython

238 - Exception in ReplWindow.cs when opening REPL window for first time
	The REPL window needs to ensure that there’s always a single span in the projection buffer, otherwise it’ll throw when something checks if it’s read-only

240 - "Stop Debugging" when attached to REPL does not detach from REPL
	Perform a detach instead of killing the process when the user asks to stop debugging

237 IronPython does not respect "wait for input when process exits" settings.
	Respect the wait for input on abnormal exit setting when we’re not debugging

241 When the interpreter is changed from IronPython to CPython, the debug launcher should be changed to Standard Python Launcher
	Warn the user when launching a CPython interpreter with the IronPython launcher.  We won’t automatically change the setting, they’ll need to do that themselves.

221 Analysis does not understand "from __future__ import unicode_literals"
	Check UnicodeLiterals on tokenizer to determine if we should create a Unicode string

222 We should recognize the magic "coding" comment (or its absence)
	Give the parser a stream when parsing the code – we get rid of our dependency on the parser needing a seekable stream by saving the bytes we read and putting them in a new stream, and then we do on-demand encoding of the buffer into bytes.  The parser finally decodes again.
	When we get a BadSourceException we should add an error for the appropriate location where we failed to decode
	When we have the BOM we should use UTF8 decoding

Also fixes an issue where we’re using the wrong root tree to resolve line number information resulting in find all refs returning incorrect locations

+++ Release/Product/Python/Debugger/Debugger/DebugEngine/AD7Engine.cs
@@ -581,7 +581,11 @@ namespace Microsoft.PythonTools.Debugger.DebugEngine {
                 return VSConstants.S_FALSE;
             }
 
-            _process.Terminate();
+            if (!_pseudoAttach) {
+                _process.Terminate();
+            } else {
+                _process.Detach();
+            }
 
             return VSConstants.S_OK;
         }
